# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RHH Property Revenue Forecaster API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /owners:
    get:
      operationId: listOwners
      tags: [owners]
      summary: List all owners
      responses:
        "200":
          description: List of owners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Owner"
    post:
      operationId: createOwner
      tags: [owners]
      summary: Create a new owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerInput"
      responses:
        "201":
          description: Created owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
  /owners/{id}:
    get:
      operationId: getOwner
      tags: [owners]
      summary: Get owner by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
    patch:
      operationId: updateOwner
      tags: [owners]
      summary: Update an owner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerUpdate"
      responses:
        "200":
          description: Updated owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
    delete:
      operationId: deleteOwner
      tags: [owners]
      summary: Archive an owner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # PROPERTIES
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a new property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"

  # MARKET AREAS
  /market/areas:
    get:
      operationId: listMarketAreas
      tags: [market]
      summary: List market areas
      responses:
        "200":
          description: Market areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketArea"
    post:
      operationId: createMarketArea
      tags: [market]
      summary: Create a market area
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketAreaInput"
      responses:
        "201":
          description: Created market area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketArea"
  /market/areas/{id}:
    patch:
      operationId: updateMarketArea
      tags: [market]
      summary: Update a market area
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketAreaUpdate"
      responses:
        "200":
          description: Updated market area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketArea"
    delete:
      operationId: deleteMarketArea
      tags: [market]
      summary: Delete market area
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # UNIT TYPE BENCHMARKS
  /market/benchmarks:
    get:
      operationId: listBenchmarks
      tags: [market]
      summary: List unit type benchmarks
      responses:
        "200":
          description: Benchmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UnitBenchmark"
    post:
      operationId: createBenchmark
      tags: [market]
      summary: Create a benchmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitBenchmarkInput"
      responses:
        "201":
          description: Created benchmark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitBenchmark"
  /market/benchmarks/{id}:
    patch:
      operationId: updateBenchmark
      tags: [market]
      summary: Update a benchmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitBenchmarkUpdate"
      responses:
        "200":
          description: Updated benchmark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitBenchmark"
    delete:
      operationId: deleteBenchmark
      tags: [market]
      summary: Delete benchmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # FORECASTS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    Owner:
      type: object
      required: [id, ownerType, firstName, lastName, email, createdAt]
      properties:
        id:
          type: number
        ownerType:
          type: string
          enum: [individual, company]
        title:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        companyName:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        preferredLanguage:
          type: ["string", "null"]
        leadSource:
          type: ["string", "null"]
        isExistingClient:
          type: boolean
        objectives:
          type: array
          items:
            type: string
        assignedToId:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    OwnerInput:
      type: object
      required: [ownerType, firstName, lastName, email]
      properties:
        ownerType:
          type: string
          enum: [individual, company]
        title:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        companyName:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        nationality:
          type: string
        preferredLanguage:
          type: string
        leadSource:
          type: string
        isExistingClient:
          type: boolean
        objectives:
          type: array
          items:
            type: string
        assignedToId:
          type: number
        notes:
          type: string
    OwnerUpdate:
      type: object
      properties:
        title:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        companyName:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        nationality:
          type: string
        preferredLanguage:
          type: string
        leadSource:
          type: string
        isExistingClient:
          type: boolean
        objectives:
          type: array
          items:
            type: string
        assignedToId:
          type: number
        notes:
          type: string

    # PROPERTIES
    Property:
      type: object
      required: [id, ownerId, emirate, area, propertyType, bedrooms, bathrooms, internalArea, createdAt]
      properties:
        id:
          type: number
        ownerId:
          type: number
        ownerName:
          type: ["string", "null"]
        emirate:
          type: string
        area:
          type: string
        community:
          type: ["string", "null"]
        development:
          type: ["string", "null"]
        projectBuilding:
          type: ["string", "null"]
        tower:
          type: ["string", "null"]
        unitNumber:
          type: ["string", "null"]
        floor:
          type: ["number", "null"]
        propertyType:
          type: string
          enum: [apartment, duplex, penthouse, townhouse, villa, studio, hotel_apartment, other]
        bedrooms:
          type: number
        bathrooms:
          type: number
        hasMaidsRoom:
          type: boolean
        hasStudy:
          type: boolean
        balconies:
          type: ["number", "null"]
        parkingSpaces:
          type: ["number", "null"]
        internalArea:
          type: number
        externalArea:
          type: ["number", "null"]
        furnishingStatus:
          type: string
          enum: [unfurnished, partially_furnished, fully_furnished, premium_furnished, hotel_grade]
        propertyCondition:
          type: string
          enum: [new, excellent, good, requires_refresh, requires_renovation]
        view:
          type: ["string", "null"]
        floorCategory:
          type: ["string", "null"]
        isWaterfront:
          type: boolean
        hasPrivatePool:
          type: boolean
        dctPermitStatus:
          type: ["string", "null"]
        currentTenancyStatus:
          type: ["string", "null"]
        currentAnnualRent:
          type: ["number", "null"]
        availabilityDate:
          type: ["string", "null"]
        heroImageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PropertyInput:
      type: object
      required: [ownerId, emirate, area, propertyType, bedrooms, bathrooms, internalArea]
      properties:
        ownerId:
          type: number
        emirate:
          type: string
        area:
          type: string
        community:
          type: string
        development:
          type: string
        projectBuilding:
          type: string
        tower:
          type: string
        unitNumber:
          type: string
        floor:
          type: number
        propertyType:
          type: string
          enum: [apartment, duplex, penthouse, townhouse, villa, studio, hotel_apartment, other]
        bedrooms:
          type: number
        bathrooms:
          type: number
        hasMaidsRoom:
          type: boolean
        hasStudy:
          type: boolean
        balconies:
          type: number
        parkingSpaces:
          type: number
        internalArea:
          type: number
        externalArea:
          type: number
        furnishingStatus:
          type: string
          enum: [unfurnished, partially_furnished, fully_furnished, premium_furnished, hotel_grade]
        propertyCondition:
          type: string
          enum: [new, excellent, good, requires_refresh, requires_renovation]
        view:
          type: string
        floorCategory:
          type: string
        isWaterfront:
          type: boolean
        hasPrivatePool:
          type: boolean
        dctPermitStatus:
          type: string
        currentTenancyStatus:
          type: string
        currentAnnualRent:
          type: number
        availabilityDate:
          type: string
        heroImageUrl:
          type: string
    PropertyUpdate:
      type: object
      properties:
        emirate:
          type: string
        area:
          type: string
        community:
          type: string
        development:
          type: string
        projectBuilding:
          type: string
        unitNumber:
          type: string
        floor:
          type: number
        propertyType:
          type: string
        bedrooms:
          type: number
        bathrooms:
          type: number
        internalArea:
          type: number
        furnishingStatus:
          type: string
        propertyCondition:
          type: string
        view:
          type: string
        dctPermitStatus:
          type: string
        currentTenancyStatus:
          type: string
        currentAnnualRent:
          type: number
        availabilityDate:
          type: string
        heroImageUrl:
          type: string

    # MARKET
    MarketArea:
      type: object
      required: [id, emirate, area]
      properties:
        id:
          type: number
        emirate:
          type: string
        area:
          type: string
        development:
          type: ["string", "null"]
        projectBuilding:
          type: ["string", "null"]
        developer:
          type: ["string", "null"]
        projectStatus:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MarketAreaInput:
      type: object
      required: [emirate, area]
      properties:
        emirate:
          type: string
        area:
          type: string
        development:
          type: string
        projectBuilding:
          type: string
        developer:
          type: string
        projectStatus:
          type: string
    MarketAreaUpdate:
      type: object
      properties:
        emirate:
          type: string
        area:
          type: string
        development:
          type: string
        projectBuilding:
          type: string
        developer:
          type: string
        projectStatus:
          type: string
        isActive:
          type: boolean
    UnitBenchmark:
      type: object
      required: [id, marketAreaId, propertyType, bedrooms]
      properties:
        id:
          type: number
        marketAreaId:
          type: number
        area:
          type: ["string", "null"]
        projectBuilding:
          type: ["string", "null"]
        propertyType:
          type: string
        bedrooms:
          type: number
        typicalAdr:
          type: ["number", "null"]
        lowSeasonAdr:
          type: ["number", "null"]
        shoulderSeasonAdr:
          type: ["number", "null"]
        peakSeasonAdr:
          type: ["number", "null"]
        eventAdr:
          type: ["number", "null"]
        expectedOccupancy:
          type: ["number", "null"]
        annualLtr:
          type: ["number", "null"]
        minLtr:
          type: ["number", "null"]
        maxLtr:
          type: ["number", "null"]
        confidenceLevel:
          type: ["string", "null"]
        sourceDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
    UnitBenchmarkInput:
      type: object
      required: [marketAreaId, propertyType, bedrooms]
      properties:
        marketAreaId:
          type: number
        propertyType:
          type: string
        bedrooms:
          type: number
        typicalAdr:
          type: number
        lowSeasonAdr:
          type: number
        shoulderSeasonAdr:
          type: number
        peakSeasonAdr:
          type: number
        eventAdr:
          type: number
        expectedOccupancy:
          type: number
        annualLtr:
          type: number
        minLtr:
          type: number
        maxLtr:
          type: number
        confidenceLevel:
          type: string
        sourceDate:
          type: string
        notes:
          type: string
    UnitBenchmarkUpdate:
      type: object
      properties:
        typicalAdr:
          type: number
        lowSeasonAdr:
          type: number
        shoulderSeasonAdr:
          type: number
        peakSeasonAdr:
          type: number
        eventAdr:
          type: number
        expectedOccupancy:
          type: number
        annualLtr:
          type: number
        minLtr:
          type: number
        maxLtr:
          type: number
        confidenceLevel:
          type: string
        notes:
          type: string
        isActive:
          type: boolean

    # FORECASTS
...