# lib/api-zod/src/generated/types/ownerUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RHH Property Revenue Forecaster API
 * OpenAPI spec version: 0.1.0
 */

export interface OwnerUpdate {
  title?: string;
  firstName?: string;
  lastName?: string;
  companyName?: string;
  email?: string;
  phone?: string;
  whatsapp?: string;
  nationality?: string;
  preferredLanguage?: string;
  leadSource?: string;
  isExistingClient?: boolean;
  objectives?: string[];
  assignedToId?: number;
  notes?: string;
}