# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RHH Property Revenue Forecaster API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyFurnishingStatus } from './propertyFurnishingStatus';
import type { PropertyPropertyCondition } from './propertyPropertyCondition';
import type { PropertyPropertyType } from './propertyPropertyType';

export interface Property {
  id: number;
  ownerId: number;
  /** @nullable */
  ownerName?: string | null;
  emirate: string;
  area: string;
  /** @nullable */
  community?: string | null;
  /** @nullable */
  development?: string | null;
  /** @nullable */
  projectBuilding?: string | null;
  /** @nullable */
  tower?: string | null;
  /** @nullable */
  unitNumber?: string | null;
  /** @nullable */
  floor?: number | null;
  propertyType: PropertyPropertyType;
  bedrooms: number;
  bathrooms: number;
  hasMaidsRoom?: boolean;
  hasStudy?: boolean;
  /** @nullable */
  balconies?: number | null;
  /** @nullable */
  parkingSpaces?: number | null;
  internalArea: number;
  /** @nullable */
  externalArea?: number | null;
  furnishingStatus?: PropertyFurnishingStatus;
  propertyCondition?: PropertyPropertyCondition;
  /** @nullable */
  view?: string | null;
  /** @nullable */
  floorCategory?: string | null;
  isWaterfront?: boolean;
  hasPrivatePool?: boolean;
  /** @nullable */
  dctPermitStatus?: string | null;
  /** @nullable */
  currentTenancyStatus?: string | null;
  /** @nullable */
  currentAnnualRent?: number | null;
  /** @nullable */
  availabilityDate?: string | null;
  /** @nullable */
  heroImageUrl?: string | null;
  createdAt: Date;
}