# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RHH Property Revenue Forecaster API
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyUpdate {
  emirate?: string;
  area?: string;
  community?: string;
  development?: string;
  projectBuilding?: string;
  unitNumber?: string;
  floor?: number;
  propertyType?: string;
  bedrooms?: number;
  bathrooms?: number;
  internalArea?: number;
  furnishingStatus?: string;
  propertyCondition?: string;
  view?: string;
  dctPermitStatus?: string;
  currentTenancyStatus?: string;
  currentAnnualRent?: number;
  availabilityDate?: string;
  heroImageUrl?: string;
}